test: move mutation tests to separate class
